Refactored setup classes
